Update README.md
Made multiple corrections and additions to the code to make all of the loops active and true when run. Collaborated with Partner to complete this project. Credit goes to Joshua Harlee for creating the main statement and disqualifier 1 and 2, Credit for the variables, For Loop and disqualifier 3 goes to me. All in all, this was a well-collaborated project and all the work was shared.